distclean: more closely follow automake's guidelines
According to the documentation, `distclean` should delete files that
`configure` built. Since many of the files we have been declaring for
`distclean-local` are generated by `autoreconf`, avoid deleting several
of those files.

Also, note that the `autom4te.cache` directory is deleted by
`maintainer-clean` already. Most of the others are required for a clean
`configure` run.

To restore the tree to a pristine state, use `git clean -xffd`.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>